Updating the Task from Theme 2

+++ basic-java/src/homework/task2/Theme2Main.java
@@ -3,11 +3,11 @@ package homework.task2;
 public class Theme2Main {
     public static void main(String[] args) {
         double price = 10.5d;
-        float ivaPercent = 16f;
+        double ivaPercent = 16d;
         double resultPrice = getPrice(price, ivaPercent);
         System.out.println(resultPrice);
     }
-    private static double getPrice(double price, float ivaPercent) {
+    private static double getPrice(double price, double ivaPercent) {
         double iva =  (ivaPercent / 100) * price;
         return iva + price;
     }
